+++ desktop/cmp/dash/DashViewSpec.js
@@ -30,8 +30,8 @@ export class DashViewSpec {
 
     /**
      * @param {string} id - unique identifier of the DashViewSpec
-     * @param {(Object|function)} content - content to be rendered by this DashView.
-     *      HoistComponent or a function returning a react element.
+     * @param {(ReactElement|Object|function)} content - content to be rendered by this DashView.
+     *      Element, HoistComponent or a function returning a react element.
      * @param {string} [title] - Title text added to the tab header.
      * @param {Element} [icon] - An icon placed at the left-side of the tab header.
      * @param {string} [groupName] - Group name to display within the add view component.